Fix two problems in import dialog: error in button responses and better support for existing directory
It appears that the previous fix for too many buttons on the import dialog did not assign actions to the buttons correctly. This should now be fixed.

In previous versions, OSCAR would present a dialog saying that it was scanning for possible CPAP data sources and then cancels the scan when it realizes that there was a previously used import directory with data in it. Now, checks for a previously used directory first and doesn't start the scan unless that directory no longer contains valid CPAP data. This should improve visual performance for the most common use case (same directory as last import).